commentaire de la partie GPS  ne fonctionnant pas encore

+++ app/src/main/java/com/example/rpgproject/rpgproject/modele/Joueur.java
@@ -94,7 +94,7 @@ public class Joueur extends Personnage {
     }
 
     @Override
-    public float getChance() {
+    public int getChance() {
         return 0;
     }
