+++ app/src/main/java/com/example/rpgproject/rpgproject/modele/Monstres/Gobelin.java
@@ -43,7 +43,7 @@ public class Gobelin extends Monstre{
     }
 
     @Override
-    public float getChance() {
+    public int getChance() {
         return 5;
     }
