+++ app/src/main/java/com/example/rpgproject/rpgproject/modele/Personnage.java
@@ -1,7 +1,5 @@
 package com.example.rpgproject.rpgproject.modele;
 
-import java.util.List;
-
 /**
  * Created by Dorian on 16/01/2015.
  */
@@ -15,7 +13,7 @@ public abstract class Personnage {
 
     public abstract int getDefense();
 
-    public abstract float getChance();
+    public abstract int getChance();
 
     public abstract String getNom();
